backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.2
djangorestframework-simplejwt>=5.2.2
psycopg2-binary>=2.9.6
gunicorn>=21.2.0
python-dotenv>=1.0.0
celery>=5.3.1
redis>=4.6.0
stripe>=6.0.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.4.1
django-stubs>=4.2.3
djangorestframework-stubs>=3.14.2
